requirements.txt
absl-py==2.1.0
ale-py==0.8.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
blinker==1.8.2
bsuite==0.3.5
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1712429964190/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
datatops==0.2.2
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1719378691416/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-acme==0.2.2
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.20
docstring_parser==0.16
etils==1.7.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1719088281970/work
Flask==3.0.3
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1720359069766/work
frozendict==2.4.4
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1719514913127/work
glfw==2.7.0
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1715527354984/work
gym==0.26.2
gym-notices==0.0.8
idna==3.7
imagecodecs @ file:///Users/runner/miniforge3/conda-bld/imagecodecs_1719236097776/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1719234999090/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1717461632069/work
immutabledict==4.2.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
importlib_resources==6.4.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1719582526268/work
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1710257588864/work
jupyterlab_widgets==3.0.11
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1695379982481/work
labmaze==1.0.6
lazy_loader==0.4
lxml==5.2.2
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1706900061465/work
matplotlib==3.9.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mizani==0.11.4
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
mujoco==3.2.0
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1715897645859/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1704469236901/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1719903567335/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
plotnine==0.13.6
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==3.20.3
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1719274593229/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==17.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyOpenGL==3.1.7
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1709721012883/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1695373498369/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1715024391267/work
requests==2.32.3
scikit-image==0.24.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1720323036795/work/dist/scipy-1.14.0-cp310-cp310-macosx_11_0_arm64.whl#sha256=b939069a5cdf606593597be906da29711202c602e3f753e5fbc9256c5a5e562d
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1714494649443/work
simple_parsing==0.1.5
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
social_agents==0.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1715941249886/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1721004824012/work
tensorflow-datasets==4.9.6
tensorflow-metadata==1.15.0
termcolor==2.4.0
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1720010439746/work
toml==0.10.2
tonic==0.3.0
torch @ file:///Users/runner/miniforge3/conda-bld/libtorch_1719361051023/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1717722719810/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1695848003431/work
urllib3==2.2.2
vibecheck==0.0.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
